Refactor matrix handling: move matrix_table function to matrix.py and update imports
