+++ libs/domain/aggregate.py
@@ -2,13 +2,12 @@
 libs/domain/aggregate.py
 """
 
-from typing import Any, Optional
+from typing import Optional
 
 import numpy as np
 import pandas as pd
 
 import libs.global_value as g
-from libs.utils import textutil
 
 
 def game_summary(
@@ -139,102 +138,3 @@ def grade_promotion_check(grade_level: int, point: int, rank: int) -> tuple[int,
             new_point = int(tbl_data[grade_level]["point"][0])  # 初期値
 
     return (new_point, grade_level)
-
-
-def matrix_table() -> pd.DataFrame:
-    """
-    対局対戦マトリックス表の作成
-
-    Returns:
-        pd.DataFrame: 集計結果
-
-    """
-    # データ収集
-    df = g.params.read_data("REPORT_MATRIX_TABLE").set_index("playtime")
-
-    # 結果に含まれるプレイヤーのリスト
-    plist = sorted(list(set(df["p1_name"].tolist() + df["p2_name"].tolist() + df["p3_name"].tolist() + df["p4_name"].tolist())))
-
-    # 順位テーブルの作成
-    l_data: dict[str, Any] = {}
-    for pname in plist:
-        if g.params.individual:  # 個人集計
-            l_name = textutil.name_replace(pname)
-            # プレイヤー指定があるなら対象以外をスキップ
-            if g.params.player_list:
-                if l_name not in g.params.player_list:
-                    continue
-            # ゲスト置換
-            if g.params.guest_skip:  # ゲストあり
-                l_name = textutil.name_replace(pname, add_mark=True)
-            else:  # ゲストなし
-                if pname == g.cfg.member.guest_name:
-                    continue
-        else:  # チーム集計
-            l_name = pname
-
-        l_data[l_name] = []
-        for x in df.itertuples():
-            match pname:
-                case x.p1_name:
-                    l_data[l_name] += [x.p1_rank]
-                case x.p2_name:
-                    l_data[l_name] += [x.p2_rank]
-                case x.p3_name:
-                    l_data[l_name] += [x.p3_rank]
-                case x.p4_name:
-                    l_data[l_name] += [x.p4_rank]
-                case _:
-                    l_data[l_name] += [None]
-
-    # 規定打数以下を足切り
-    if g.params.stipulated:
-        for pname in list(l_data.keys()):
-            if sum(x is not None for x in l_data[pname]) < g.params.stipulated:
-                l_data.pop(pname)
-
-    rank_df = pd.DataFrame(l_data.values(), columns=list(df.index), index=list(l_data.keys()))
-
-    # 対象リストが0件になった場合は空のデータフレームを返す
-    if rank_df.empty:
-        return rank_df
-
-    # 対局対戦マトリックス表の作成
-    mtx_df = pd.DataFrame(index=list(l_data.keys()), columns=list(l_data.keys()) + ["total"])
-    sorting_df = pd.DataFrame(index=list(l_data.keys()), columns=["win_per", "count"])
-
-    for idx1 in range(len(rank_df)):
-        p1 = rank_df.iloc[idx1]
-        t_game_count = 0
-        t_win = 0
-        for idx2 in range(len(rank_df)):
-            p2 = rank_df.iloc[idx2]
-            if p1.name == p2.name:
-                mtx_df.loc[f"{p1.name}", f"{p2.name}"] = "---"
-            else:
-                game_count = len(pd.concat([p1, p2], axis=1).dropna())
-                win = (p1 < p2).sum()
-                t_game_count += game_count
-                t_win += win
-
-                if game_count:
-                    winning_per = str(round(float(win / game_count * 100), 1))
-                else:
-                    winning_per = "--.-"
-                mtx_df.loc[f"{p1.name}", f"{p2.name}"] = f"{win}-{game_count - win} ({winning_per}%)"
-
-        if t_game_count:
-            t_winning_per = str(round(float(t_win / t_game_count * 100), 1))
-        else:
-            t_winning_per = "--.-"
-        mtx_df.loc[f"{p1.name}", "total"] = f"{t_win}-{t_game_count - t_win} ({t_winning_per}%)"
-        sorting_df.loc[f"{p1.name}", "win_per"] = t_winning_per
-        sorting_df.loc[f"{p1.name}", "count"] = t_game_count
-
-    # 勝率で並び替え
-    sorting_df["win_per"] = pd.to_numeric(sorting_df["win_per"], errors="coerce")
-    sorting_df["count"] = pd.to_numeric(sorting_df["count"], errors="coerce")
-    sorting_df = sorting_df.sort_values(by=["win_per", "count"], ascending=[False, False])
-    mtx_df = mtx_df.reindex(index=list(sorting_df.index), columns=list(sorting_df.index) + ["total"])
-
-    return mtx_df
